Add files via upload
This is a set of programmed named with respect to the order of questions given. The order_sender.py runs the node for sending orsers to the robot while the confirmation.py runs the nonde for giving confirmation to the robot from the table and kitchen. CancelOrder.srv holds a boolean variable for cancelling orders.